pojo registrar init
